MODIFY CSS FOR ABOUT PAGE AND USER LOGIN HEADER TO MAKE MOBILE FRIENDLY
Tweaked Css on the about page to add flexbox for the rules box. Logic on mouseover highlights
relevant rules to the icon and the box will flex-wrap on small screens to keep it on
screen.
Also made changes to opacity of the logged in user data so if scrolled over
will still show text below - relevant on samaller screens.